requirements.txt
streamlit
python-dotenv
openai
llama-index
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
llama-index-vector-stores-chroma
chromadb
pandas